Add functions for additional sytem information
This will eventually be used to improve the output when running
benchmarks in Benchee. For now, though, this is pretty much just the
rough outline of how to get this information so I can get some feedback
on if I'm on the right path according to what @PragTob was envisioning.

Eventually this will close #63.