Refactor implementation
Originally I had implement all of the platforms in one go, but I think
@PragTob's idea of doing more, smaller PRs is a good one. This
refactoring cleans up the partially finished work I had done so it's
ready to be merged.

There is potentially more refactoring that can be done, but I think it
will be helpful to see what is output from the other system calls that
we need to do for the other platforms (Linux and Windows) before going
down that path.