Really rough implementation of calculating memory usage
This adds in a non-breaking way to calculate memory usage. Right
now the actual calculation of the memory usage info is probably the
least good thing about this commit, since I'm seeing wildly varying
results (including negative memory usage) because of the inability to
turn off the GC. However, this is implemented in a way that changing
that calculation later on shouldn't require any further changes to what
we have here.

Next up is to do something with the results in the UI, and once we have
something rough there we can do some touchup work on this really naieve
implementation.