WIP - Begin refactoring of Benchmark module
This is a non-functional, very eary, broken step in the process. I'm
just putting it up now to get some feedback before continuing.

While I was working on adding the memory usage information, it struck
me that it would be a lot easier to add that without major changes if
there was already a struct representing an individual benchmark that I
could add it to. I decided to explore this refactoring a bit, and after
this initial step I kind of like it.

Basically, I split the `Benchee.Benchmark` module into two parts -
`Benchee.Benchmark`, which handles creating the benchmark structs and
passing them along to the actual functions that will run the benchmarks,
and another module (`Benchee.Benchmark.Runner`) that handles the actual
running of the benchmarks.

So, what do folks think? It made sense to me since we already had a
struct for a `Suite` of benchmarks.