Change file name for `Benchee.Utility.MapValues`
I noticed while I was trying to goto file for this module that it
couldn't find the file in the path, and there was a slight discrepency
between the module name and the file name. This is a tiny change, but it
seems like it might make life a little easier just to follow the
conventions so tooling all works as it should.