Add Benchmark.Scenario struct
This commit introduces a new struct representing a single thing to
benchmark, which we're calling a Scenario. A scenario represents a
combination of a function and an optional input (or inputs) for that
function. Right now the creation of these scenarios for each job and
each input is implemented and tested, but we're not actually using them
yet.