Refactor Benchee.Benchmark and tests
This commit has Benchee.Benchmark refactored and in good shape
(including documentation) after extracting the new runner from that
module. The implementation at this point is still pretty much broken,
but next is to get the runner up to shape and tested. Once the runner
is implemented and tested, then we'll need to update the statistics.