Re-add parallel calculation of statistics
Mea culpa on this one. I remember removing that and thinking "I'll add
this back in once the tests are green", but then I never did! I wonder,
how might we add a test to ensure that we're always calculating these
in parallel? We can't rely on the same trick we did to ensure that
scenarios are run in parallel, but I'm sure we could come up with
something.

I also noticed now that the old `MapValues` module is no longer used,
so I've removed that in this commit.