Extract `Benchee.Utility.Paralell` module with `map/1`
This extracts the parallel mapping function that was present in both
`Benchmark.Runner` and `Benchmark.Statistics`. I had to modify the
function that we're passing in as an argument in `Runner` to accept an
argument that is just thrown away. I hope this isn't too confusing, but
it's the best way I could think of making this work.