Quick spike on refactor idea for `System` module
This is just a first quick spike on an idea for refactoring the
`System` module so we can better test edge cases with some of these
interactions with the scary world outside our system. Instead of
injecting `System.system_cmd/3` into all the functions that use it
(which was a lot), I figured we might be able to pass a success
callback to `System.system_cmd/4` and test these edge cases at that
level.

It would totally work the other way, too, and I might do that tomorrow
as well just for comparison's sake, but I figured I might put this up
now to spur conversation. I think I liked the idea in my head more than
I like this actual implementation, but it's done now so I think it's
worth a push up to get feedback from others.