Interpret system returned memory size as binary
This always irked, I think a binary representation is correct as
you know, computers.

The wikipedia article (https://en.wikipedia.org/wiki/Gigabyte)
also claims it:

However, the term is also used in some fields of computer science
and information technology to denote 1073741824 (1024^3 or 2^30)
bytes, particularly for sizes of RAM.

So, I think this is more correct. I also noted that we don't test
that particular part, but that might be for another day :)