Implement mode in Statistics calculation
Another nice statistics to have handy. Imo could be very useful
for memory measuremnts as I hope that memory usage will often be
very stable so getting the mode should be close to the actual
memory usage.

We also get the `mode_count` for "free` not sure if we should
include it or not (I have never seen it used), but it seems sort
of useful (like for memory usages we could say/include stuff like
60% of all samples were this that helps)