Define behaviour for a formatter and output in parallel
In this commit I'm defining a behaviour for a Benchee Formatter, which
consistes of three functions:

* format/1 which takes a Benchee.Suite struct and formats results
* write/1 which takes the output from format/1 and writes it either to
a file or to the console, and
* output/1 which basically just calls `format/1` and then `write/1` and
returns the Benchee.Suite struct it was given.

I don't see why we can't eventually remove the `output/1` callback from
this behaviour once we've switched over to using `parallel_output/2`
and once we've implemented this behaviour in all the other formatters.

I also wrote a function that can do as much in parallel as possible for
formatting. We can't write the output in parallel, but we can generate
the output in parallel, so we're doing that here.

I haven't switched over to actually using this new function to do the
formatting in parallel, but we can do that in a later commit.

Resolves #53