Make sure overhead for fast functions is as slim as possible
Special case the case when fast functions don't have before_each/
after_each so that the performance of running very fast functions
is back on par with the performance of before the after_each/
before_each cases.

Performance difference was significant, as in without this special
case super fast functions got a reported average of 0.082 microseconds
and now they are down to 0.014 which is on par with current master.

(in my testing even a tiny bit faster)