before_scenario is passed the input and can return a different one
Introduced a new `scenario_input` field to keep the value around
as long as we need it as the hooks are run exactly once for each
scenario. Contemplated just naming it `input` but decided it was
too difficult.

As a side effect, creation of the benchmarking function had to be
moved around. It's closer to where it will be used now, which is
good. However, once we pass this on to `before/after_each` there
is a strong possibility it might change as well (as the inputs can
change again).