before_each gets inputs and passes them on
Sadly, this breaks our new found great achievement that we have
a benchmarking function that we can reuse. The reason being, that
when before_each can change the input for every invocation - so
the function we create has to be recreated every time.

This could potentially be optimized further, by using `:timer.tc/2`
or be determining if before_* methods are given at all. Plus for
our "repeat n" case, the function could be reused and doesn't have
to be recreated everytime.