Add percentiles to statistics
Also adds the ability to pass a list of percentile ranks into
`percentile/2` to calculate multiple percentiles without sorting
the list multiple times.

This form is used currently in `Statistics`, so we only sort the list
once and then calculate the 50th and 99th percentiles (where 50th is the
median).