Provide .tool-versions with current go do elixir/erlang versions
Interested to see how people feel about this.

a) Not everyone uses asdf so not directly affected
b) if you know where to look (travis) you already have a place to
get an overview of supported versions. Should also be in the
README but is probably missing atm ;P
c) potentially high churn (ideally it'll be updated rather quickly
for every new elixr/erlang release)

So, does it provide value - yay or nay?