use Formatter to avoid extra logic in `output/1` + repetition
Noticed it in the HTML formatter that the output/1 function had
added logic which then wouldn't be called in the new "specify
a formatter as a module and we'll call `format/1` and `output/1`
for you". `output/1` should really only do this, nothing more.

This helps us enforce this and avoid code duplication.