Add experiment with controlling garbage collection
It turns out that `Process.spawn/2` already accepted the options that
`erlang.spawn_opt/4` could take, so we could use that to set up our
process for us.

This is a super bare bones implementation with a very simple
supervision tree setup. Our main process just watches for `:DOWN`
messages from the tasks and exits if that happens. We should probably
do something else to make this a better experience for users in the
event that they do crash a process with their benchmarking scenarios.