Refactor where we're running things in parallel
The original fix I had for this was kind of hacky, but this moves the
parallelization down to just where we're measuring runtimes. This
should actually be slightly more efficient when we're running
benchmarks in parallel (and less buggy) since before we were running
the `before_scenario` and `after_scenario` hooks before each _instance_
of the benchmark (when running in parallel). Now we run them only once
before each scenario, and each scenario itself is run in parallel.