Add measurement of memory usage
This gives us a very basic measurement of memory usage. It is
constrained to measuring the memory used by the _single_ process
executing the function. If the function being measured spawns further
processes, the memory used by those processes would not be counted with
this method. However, this single process measurement represents the
vast majority of use cases for Benchee, so it's a solid first step.