Format Benchee.BenchmarkTest
We've got a lot of formatting to do. Might as well start with some of
our low churn code! I was thinking we'd take it file by file like they
did with the Elixir codebase itself instead of trying to do one huge
piece.

Once we have everything formatted then we can add a formatter check to
CI so we stay that way.