First version of separating memory and runtime measurements
Introduces a new `Measurer` behaviour that can be adopted by `Time`
and `Memory` and in the future `Reduction`. Also switches over to
memory having its own configuration for time. Like how this ended up
in general.

This isn't merge ready yet, it's a first implementation already
up for review and CI.

TODO:
- [ ] README/documentation adjustments
- [ ] respect `memory_time` during calculation of estimated total runtime
- [ ] allow measurer to return `nil` when a a measurement "failed" and handle it appropriately (memory measurement is less than 0 case)