Handle calculation of statistics when given 0s
We hadn't handled the calculation of statistics when given a whole
bunch of 0s, which is a complication since we're doing a lot of
division in there. Technically if we're in a situation where we're
supposed to divide by 0 the answer is Infinity, but instead of
introducing that, for now I'm just returning 0 as well.

Addresses the first reported bug in #213.