Memory breaker integration test
Follow up to #213

Decided to inline the relevant parts of it to the test suite as
it highlighted so many problems. One of which I don't know how
to test. Moreover, with it highlighting so many problems it might
highlight them again later.

Also removed some dead code from said benchmark and added
attribution to @michalmuskala