Fix race condition in spawning tracer process
We were somewhat reliably missing the first garbage collection event
because when we spawned our tracer, we were setting the tracing flag in
the function to be executed in our tracer. This means the runner and
tracer were running concurrently for a moment, and the runner hit a GC
threshold before the tracer was able to set the appropriate tracing
flag!

Now we explicitly block the runner from running the function to be run
until the tracing flag is set.