Re-write memory collection to be more accurate
This removes the `Process.info/1` check around the function to be
benchmarked to see what's still on the heap and relies solely on
garbage collection for our measurements. I had to make substantial
changes to the way we're collecting memory - including being much more
specific about turning on and off the collection of messages - but it's
yielding much more accurate results - and it works on OTP 21!