Fix race condition in spawning tracer process (#239)
* Fix race condition in spawning tracer process

We were somewhat reliably missing the first garbage collection event
because when we spawned our tracer, we were setting the tracing flag in
the function to be executed in our tracer. This means the runner and
tracer were running concurrently for a moment, and the runner hit a GC
threshold before the tracer was able to set the appropriate tracing
flag!

Now we explicitly block the runner from running the function to be run
until the tracing flag is set.

* Re-write memory collection to be more accurate

This removes the `Process.info/1` check around the function to be
benchmarked to see what's still on the heap and relies solely on
garbage collection for our measurements. I had to make substantial
changes to the way we're collecting memory - including being much more
specific about turning on and off the collection of messages - but it's
yielding much more accurate results - and it works on OTP 21!

* Make it better thanks to feedback from Tobi