Remove any bindings to any data allocated during measurement
This will ensure that we're collecting _only_ data that is allocated as
part of the execution of the function we're trying to benchmark. The
downside is that we end up running the function twice - once to
capture the return value and once to measure the memory usage - but
this way we're sure that we're not accidentally counting terms that
weren't directly a result of the function we're measuring.