Fix issue of passing nil to formatters (from CSV PR)
Interestingly the bug occurred thanks to our legacy format support.
Can't wait to get rid off it ;)

I changed the code for the formatter test to really make sure that
the same options passed to format/2 are also passed to write/2.
It didn't find the bug as the bug was in Configuration, but I
think it's still an improvement worth keeping.

Once this code is merged, we can delet the one unnecessary line
from the CSV formatter. Already tried it and it worked :tada: