Allow inputs to be an ordered list of tuples
Now if inputs are given as an ordered list of tuples, the output will be
in the same order as the list of inputs given instead of sorted by the
input alphabetically.

I also updated some dependencies to get them on the same page as the
formatters, and fixed some Credo issues.