Promote Scenario to be a top level data structure
Over time Scenario has become one of our most used data structures,
perhaps the most used data structure and the most important one
besides Suite.
I mean of course, because all the important data (as in results)
is kept within a scenario. Even CollectionData is top level.
Makes no sense for Scenario to be scoped under Benchmark, as some
of the changed files also imply (as often it was the only aliased
module that was not under the top level name space)