Add reduction counting (#307)
* Add reduction collector

This adds a module that collects reduction counts for a function, and a
test for that measurement. Right now it can't really be used anywhere,
but we'll put that in in the next commit.

* Add console formatting for reductions

This now actually shows the information for reduction counting. Since
it's usually very much like the memory measurements (in that it's
consistent unless there's non-determinism in your code) I basically
copied a lot of the behavior from there.

* updating based on feedback

* Bump to newest Elixir & Erlang, and make function private

* move error handling into the console formatter