Deactivate function call overhead measurement by default
Especially with errors during measurement accuracy on certain
OS'es (hey there Mac OS and Windows) this feature could be
extremely detrimental. It'd falsely report 0 times as its
measurement wasn't precise enough so it couldn't differentiate
between a fast function and the empty function.

I'm more convinced now, that this is an advanced feature and
should hence be opt-in and not opt out.