Fix relative statistics calculations
It used to rely on the inputs being there and then working based
off on them - too bad, if we use `Benchee.load` then there are
no inputs to go off, but they are still in the scenarios to work
with them. Needs some care to preserve order which looks odd but
couldn't find a better way.