Make printer correctly handle floats and integers
Noticed this while working on something else.

There is a question on if the printer should handle this kind
of logic at all - and it's a good question. However, I didn't
want to remodel the code right now too much and the individual
measurements already do the 0 check. Yes we could avoid spawning
a process but it's an edge case so oh well.